+++ commons/jasypt-config/src/main/java/com/flab/bigtrader/JasyptConfig.java
@@ -1,26 +1,24 @@
-package com.flab.bigtrader.common.config;
+package com.flab.bigtrader;
 
 import org.jasypt.encryption.StringEncryptor;
 import org.jasypt.encryption.pbe.PooledPBEStringEncryptor;
 import org.jasypt.encryption.pbe.config.SimpleStringPBEConfig;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class JasyptConfig {
+	private final JasyptProperties jasyptProperties;
 
-	private final String password;
-
-	public JasyptConfig(@Value("${jasypt.encryptor.password}") String password) {
-		this.password = password;
+	public JasyptConfig(JasyptProperties jasyptProperties) {
+		this.jasyptProperties = jasyptProperties;
 	}
 
 	@Bean("jasyptStringEncryptor")
 	public StringEncryptor stringEncryptor() {
 		PooledPBEStringEncryptor encryptor = new PooledPBEStringEncryptor();
 		SimpleStringPBEConfig config = new SimpleStringPBEConfig();
-		config.setPassword(password);
+		config.setPassword(jasyptProperties.getPassword());
 		config.setAlgorithm("PBEWithMD5AndDES");
 		config.setKeyObtentionIterations("1000");
 		config.setPoolSize("1");
